requirements.txt
lotrek-django-hvad>=2.0.6
djsuperadmin>=0.9,<1.0.0
djangorestframework>=3.10.0<4.0.0
djlotrek>=0.1.6,<1.0.0
Pillow>=6.2.0,<7.0.0
django-ckeditor>=5.7.1,<6.0.0
python-magic>=0.4<0.5
Django>=1.8,<5